pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-regression-tester"
version = "0.1.0"
description = "A flexible library for testing LLM responses against predefined rubrics using OpenAI's API for automated scoring"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "LLM Regression Tester Contributors"}
]
maintainers = [
    {name = "LLM Regression Tester Contributors"}
]
keywords = ["llm", "testing", "evaluation", "rubrics", "ai", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "typing-extensions>=4.0.0",
    "openai>=1.0.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
# No optional dependencies - all required dependencies are now in the main dependencies list

[project.urls]
Homepage = "https://github.com/ktech99/llm-regression-tester"
Documentation = "https://llm-regression-tester.readthedocs.io/"
Repository = "https://github.com/ktech99/llm-regression-tester.git"
Issues = "https://github.com/ktech99/llm-regression-tester/issues"
Changelog = "https://github.com/ktech99/llm-regression-tester/blob/main/CHANGELOG.md"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"llm_regression_tester" = ["*.json", "*.md"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
